+++ crates/ruma-federation-api/CHANGELOG.md
@@ -5,7 +5,7 @@ Breaking changes:
 - Remove the `origin` field in `create_join_event::{v1/v2}::RoomState` due to a
   clarification in the spec.
 - The type of `signed` in `thirdparty::bind_callback::v1::Request` was fixed. It
-  uses `SignedContent` from `RoomMemberEventContent`.
+  uses `Raw<SignedContent>` from `RoomMemberEventContent`.
 - The type of `content` in `thirdparty::exchange_invite::v1::Request` was fixed.
   It is a `RoomMemberEventContent`. A new constructor was added,
   `with_third_party_invite()` that constructs the event content from a